refs #140: Moved in memory implementation to own module to help consumers test
