+++ extensions/spring-cloud-schema-registry-extension/avro-spring-cloud-schema-registry-extension/pom.xml
@@ -74,6 +74,13 @@
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
+
+        <dependency>
+            <groupId>com.jashmore</groupId>
+            <artifactId>in-memory-spring-cloud-schema-registry</artifactId>
+            <version>${project.version}</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
